refactor connector
